+++ pyo3-speedup/raytracer_rs/src/object_parser.rs
@@ -8,24 +8,24 @@ use crate::scene::*;
 pub fn _parse_vector(py_vec: PyObject, py: &Python) -> Vector{
     let object_ref: &PyAny = py_vec.extract(*py).unwrap();
     Vector{
-        x: object_ref.getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("y").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("z").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("y").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("z").unwrap().extract::<f64>().unwrap(),
     }
 }
 
 pub fn _parse_light(py_light: &PyObject, py: &Python) -> Light{
     let object_ref: &PyAny = py_light.extract(*py).unwrap();
     let v = Vector{
-        x: object_ref.getattr("position").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("position").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("position").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("position").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("position").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("position").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
     };
     let c = Color{
-        r: object_ref.getattr("color").unwrap().getattr("r").unwrap().extract::<f32>().unwrap(),
-        g: object_ref.getattr("color").unwrap().getattr("g").unwrap().extract::<f32>().unwrap(),
-        b: object_ref.getattr("color").unwrap().getattr("b").unwrap().extract::<f32>().unwrap(),
-        special: object_ref.getattr("color").unwrap().getattr("special").unwrap().extract::<f32>().unwrap(),
+        r: object_ref.getattr("color").unwrap().getattr("r").unwrap().extract::<f64>().unwrap(),
+        g: object_ref.getattr("color").unwrap().getattr("g").unwrap().extract::<f64>().unwrap(),
+        b: object_ref.getattr("color").unwrap().getattr("b").unwrap().extract::<f64>().unwrap(),
+        special: object_ref.getattr("color").unwrap().getattr("special").unwrap().extract::<f64>().unwrap(),
     };
     Light{
         position: v,
@@ -36,37 +36,37 @@ pub fn _parse_light(py_light: &PyObject, py: &Python) -> Light{
 fn _parse_triangle(py_object: &PyObject, py: &Python) -> Triangle{
     let object_ref: &PyAny = py_object.extract(*py).unwrap();
     let a = Vector{
-        x: object_ref.getattr("a").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("a").unwrap().getattr("y").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("a").unwrap().getattr("z").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("a").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("a").unwrap().getattr("y").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("a").unwrap().getattr("z").unwrap().extract::<f64>().unwrap(),
     };
     let b = Vector{
-        x: object_ref.getattr("b").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("b").unwrap().getattr("y").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("b").unwrap().getattr("z").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("b").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("b").unwrap().getattr("y").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("b").unwrap().getattr("z").unwrap().extract::<f64>().unwrap(),
     };
     let c = Vector{
-        x: object_ref.getattr("c").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("c").unwrap().getattr("y").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("c").unwrap().getattr("z").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("c").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("c").unwrap().getattr("y").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("c").unwrap().getattr("z").unwrap().extract::<f64>().unwrap(),
     };
 
     let ca = &c - &a;
     let ba = &b - &a;
 
     let normal = Vector{
-        x: object_ref.getattr("normal").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("normal").unwrap().getattr("y").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("normal").unwrap().getattr("z").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("normal").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("normal").unwrap().getattr("y").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("normal").unwrap().getattr("z").unwrap().extract::<f64>().unwrap(),
     };
 
-    let distance = object_ref.getattr("distance").unwrap().extract::<f32>().unwrap();
+    let distance = object_ref.getattr("distance").unwrap().extract::<f64>().unwrap();
 
     let color = Color{
-        r: object_ref.getattr("color").unwrap().getattr("r").unwrap().extract::<f32>().unwrap(),
-        g: object_ref.getattr("color").unwrap().getattr("g").unwrap().extract::<f32>().unwrap(),
-        b: object_ref.getattr("color").unwrap().getattr("b").unwrap().extract::<f32>().unwrap(),
-        special: object_ref.getattr("color").unwrap().getattr("special").unwrap().extract::<f32>().unwrap(),
+        r: object_ref.getattr("color").unwrap().getattr("r").unwrap().extract::<f64>().unwrap(),
+        g: object_ref.getattr("color").unwrap().getattr("g").unwrap().extract::<f64>().unwrap(),
+        b: object_ref.getattr("color").unwrap().getattr("b").unwrap().extract::<f64>().unwrap(),
+        special: object_ref.getattr("color").unwrap().getattr("special").unwrap().extract::<f64>().unwrap(),
     };
 
     Triangle{
@@ -85,18 +85,18 @@ fn _parse_triangle(py_object: &PyObject, py: &Python) -> Triangle{
 fn _parse_plane(py_object: &PyObject, py: &Python) -> Plane{
     let object_ref: &PyAny = py_object.extract(*py).unwrap();
     let normal = Vector{
-        x: object_ref.getattr("normal").unwrap().getattr("x").unwrap().extract::<f32>().unwrap(),
-        y: object_ref.getattr("normal").unwrap().getattr("y").unwrap().extract::<f32>().unwrap(),
-        z: object_ref.getattr("normal").unwrap().getattr("z").unwrap().extract::<f32>().unwrap(),
+        x: object_ref.getattr("normal").unwrap().getattr("x").unwrap().extract::<f64>().unwrap(),
+        y: object_ref.getattr("normal").unwrap().getattr("y").unwrap().extract::<f64>().unwrap(),
+        z: object_ref.getattr("normal").unwrap().getattr("z").unwrap().extract::<f64>().unwrap(),
     };
 
-    let distance = object_ref.getattr("distance").unwrap().extract::<f32>().unwrap();
+    let distance = object_ref.getattr("distance").unwrap().extract::<f64>().unwrap();
 
     let color = Color{
-        r: object_ref.getattr("color").unwrap().getattr("r").unwrap().extract::<f32>().unwrap(),
-        g: object_ref.getattr("color").unwrap().getattr("g").unwrap().extract::<f32>().unwrap(),
-        b: object_ref.getattr("color").unwrap().getattr("b").unwrap().extract::<f32>().unwrap(),
-        special: object_ref.getattr("color").unwrap().getattr("special").unwrap().extract::<f32>().unwrap(),
+        r: object_ref.getattr("color").unwrap().getattr("r").unwrap().extract::<f64>().unwrap(),
+        g: object_ref.getattr("color").unwrap().getattr("g").unwrap().extract::<f64>().unwrap(),
+        b: object_ref.getattr("color").unwrap().getattr("b").unwrap().extract::<f64>().unwrap(),
+        special: object_ref.getattr("color").unwrap().getattr("special").unwrap().extract::<f64>().unwrap(),
     };
 
     Plane{
